==> R/06_heatmap.qmd <==
---
title: "06 - Create heatmap analysis"
subtitle: "R for Bio Data Science"
format: html
author: Thomas Tams (s204540)
date: "2023-11-21"
editor: visual
execute:
  message: false
  warning: false
---

## Load libraries

```{r}
library("tidyverse")
library("here")
library("ggdendroplot")
source(here("R/99_proj_func.R"))
```

## Load selected transcript data

```{r}
select_transcript_meta <-
  read_tsv(here("data/03_dat_augment.tsv"))

select_transcript_meta
```

## Select expression data and pivot to have transcripts as rows and log2(1+FPKM) expression values

```{r}
selected_transcript_expr_log <- select_transcript_meta |>
  select(1, 34:ncol(select_transcript_meta)) |>
  pivot_longer(
    col = -sample_id,
    names_to = "transcript_id",
    values_to = "expression_value"
  ) |>
  mutate(expression_value = log2(1 + expression_value)) |>
  pivot_wider(
    names_from = sample_id,
    values_from = expression_value
  )

selected_transcript_expr_log
```

## Hierarchical clustering for Heatmap and Dendrograms

```{r}
# Put to matrix format for compatible hclust(dist())
expr_matrix <- selected_transcript_expr_log |>
  select(-transcript_id) |>
  as.matrix()

rownames(expr_matrix) <- selected_transcript_expr_log |>
  select(transcript_id) |>
  as_vector()

# Perform hierarchical clustering
transcript_clus <- expr_matrix |>
  dist() |>
  hclust()

sample_clus <- expr_matrix |>
  t() |>
  dist() |>
  hclust()

# bring the data.frame into a from easily usable by ggplot
hm <- hmReady(expr_matrix, colclus = sample_clus, rowclus = transcript_clus)

# Setup color vector for beautiful colors
color_vector <- c(
  "#5154A5", "#5154A5", "#5154A5",
  "#41B6E8", "#41B6E8",
  "#69BD44", "#69BD44", "#69BD44",
  "#FCE349", "#FCE349", "#FCE349", "#FCE349", "#FCE349", "#FCE349", "#FCE349", "#FCE349"
)

max_expr_val <- selected_transcript_expr_log |>
  group_by(transcript_id) |>
  nest() |>
  mutate(max_val_transcript = unlist(map(
    .x = data,
    .f = ~ max(.x)
  ))) |>
  pull(max_val_transcript) |>
  max()

# plot the heatmap
hmplot <- ggplot() +
  geom_tile(data = hm, aes(x = x, y = y, fill = value)) +
  scale_y_reverse() +
  geom_dendro(sample_clus, ylim = c(0, -40), axis.labels = FALSE) +
  geom_dendro(transcript_clus, xlim = c(0, -30), pointing = "side", axis.labels = FALSE) +
  scale_fill_gradientn(
    colors = color_vector,
    breaks = c(0, 7, max_expr_val),
    labels = scales::label_number(accuracy = 0.01)
  ) +
  labs(
    fill = "Log2(1 + FPKM)",
    title = "Heatmap of expression values",
    subtitle = "Hierarchical clustering of samples (x-axis) \nand transcript (y-axis)",
    caption = "Data from DOI: https://doi.org/10.1038/nature14664"
  ) +
  theme_minimal() +
  theme(
    panel.grid = element_blank(),
    axis.line = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank(),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    text = element_text(size = 20)
  )

ggsave(here("results/06_heatmap_plot.png"), hmplot, height = 10, width = 7.5)

# Printing plot
hmplot
```
