==> R/07_logistic_model.qmd <==
---
title: "07 - Logistic modelling of gene expression"
subtitle: "R for Bio Data Science"
format: html
author: Mikkel Niklas Rasmussen (s193518)
date: "2023-11-20"
editor: visual
---

## Load libraries

```{r}
library("tidyverse")
library("here")
library("broom")
library("ggrepel")
```

## Load data

```{r}
data <- read_tsv(here("data/03_dat_augment.tsv"))
```

## Drop samples with NA in cluster_sample

```{r}
data <- data |>
  drop_na(sample_cluster)
```

## Get expression data from all data

```{r}
# Extract the expression data
expression_data <- data |>
  select(sample_id, sample_cluster, 33:ncol(data))
```

## Extract genes of interest from PCA analysis

```{r}
# Subset data to expression data
input_pca <- data |>
  select(-c(1:33))

# Run pca
pca_fit <- log2(input_pca + 1) |>
  prcomp()
```

```{r}
# Convert expression data into long format
expr_long <- expression_data |>
  pivot_longer(
    cols = -c(sample_id, sample_cluster),
    names_to = "gene",
    values_to = "expr_level"
  )

# Log2 transform data
expr_log_long <- expr_long |>
  mutate(
    log2_expr_level = log2(expr_level + 1)
  )
```

## Explore cluster 1 vs all

```{r}
# Mutate binary cluster x or not
cluster_of_interest <- "Cluster_1"
expr_log_long_clust <- expr_log_long |>
  mutate(
    cluster = case_when(
      sample_cluster == cluster_of_interest ~ 1,
      sample_cluster != cluster_of_interest ~ 0
    ),
    .after = sample_id
  ) |>
  select(!sample_cluster)

expr_log_long_clust
```

```{r}
# Nest the data
expr_log_long_nested <- expr_log_long_clust |>
  group_by(gene) |>
  nest() |>
  ungroup()

# Model each gene with a logistic regression model
expr_log_long_nested <- expr_log_long_nested |>
  group_by(gene) |>
  mutate(model_object = map(
    .x = data,
    .f = ~ lm(
      formula = log2_expr_level ~ cluster,
      data = .x
    )
  ))

# Calculate confidence intervals
expr_log_long_nested <- expr_log_long_nested |>
  mutate(model_object_tidy = map(
    .x = model_object,
    .f = ~ tidy(
      x = .x,
      conf.int = TRUE,
      conf.level = 0.95
    )
  ))

# Unnest the data
expr_estimates <- expr_log_long_nested |>
  unnest(model_object_tidy) |>
  drop_na(p.value)

# Extract columns of
expr_estimates <- expr_estimates |>
  filter(term == "cluster") |>
  select(gene, p.value, estimate, conf.low, conf.high) |>
  ungroup()

expr_estimates

expr_estimates <- expr_estimates |>
  mutate(
    q.value = p.adjust(p.value),
    is_significant = case_when(
      q.value <= 0.05 ~ "yes",
      q.value > 0.05 ~ "no"
    )
  )

expr_estimates |>
  filter(is_significant == "yes") |>
  select(gene, q.value)
```

```{r}
# Select 4 transcripts based on min and max values in
# Select transcripts differentiating samples along PC1
pca_1_features <- pca_fit |>
  tidy(matrix = "rotation") |>
  pivot_wider(names_from = "PC", names_prefix = "PC", values_from = "value") |>
  select(column, PC1, PC2) |>
  arrange(desc(abs(PC1))) |>
  slice(1:13) |>
  pull(column)

# Select genes investigated in PCA
select_genes <- data |>
  select(sample_cluster, pca_1_features[1:3], pca_1_features[13]) |>
  colnames()

pl1 <- expr_estimates |>
  filter(is_significant == "yes") |>
  mutate(
    lbl = case_when(
      gene %in% select_genes ~ "prior",
      .default = is_significant
    ),
    gene = str_remove_all(gene, "_NM_.*"),
  ) |>
  group_by(gene, lbl) |>
  summarise(
    estimate = mean(estimate),
    conf.low = mean(conf.low),
    conf.high = mean(conf.high)
  ) |>
  ggplot(aes(
    x = estimate,
    y = fct_reorder(gene, estimate),
    xmin = conf.low,
    xmax = conf.high,
    color = lbl
  )) +
  geom_vline(xintercept = 0) +
  geom_errorbarh() +
  geom_point() +
  theme_minimal(base_size = 10) +
  labs(
    x = "Estimates (95%CIs)",
    y = "",
    title = "Genes Associated with Cluster 1",
    caption = "Data from DOI:https://doi.org/10.1038/nature14664"
  ) +
  theme(
    text = element_text(size = 18),
    plot.title = element_text(28),
    legend.position = "none"
  ) +
  scale_color_manual(values = c("green", "red"))
pl1
ggsave(here("results/07_associated_genes_CI.png"), pl1, height = 16, width = 10)


pl2 <- expr_estimates |>
  mutate(
    is_significant = case_when(
      gene %in% select_genes ~ "prior",
      .default = is_significant
    )
  ) |>
  mutate(
    gene = str_remove_all(gene, "_NM_.*"),
    lbl = case_when(
      is_significant == "yes" ~ gene,
      is_significant == "prior" ~ gene,
      is_significant == "no" ~ ""
    )
  ) |>
  ggplot(aes(
    x = estimate,
    y = -log10(q.value),
    colour = is_significant,
    label = lbl
  )) +
  geom_point(
    size = 0.5,
    alpha = 0.5
  ) +
  geom_text_repel(
    size = 6,
    max.overlaps = 20
  ) +
  geom_hline(yintercept = 0) +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none",
    text = element_text(size = 25)
  ) +
  labs(
    x = "Estimates",
    y = bquote(-log[10](q)),
    title = "Trancripts Associated with Patient Cluster 1",
    subtitle = str_c("Trancripts highlighted in red had a p-adj. < 0.05 \n- green were also found in PCA"),
    caption = "Data from DOI: https://doi.org/10.1038/nature14664"
  ) +
  scale_color_manual(values = c("black", "green", "red"))
pl2

ggsave(here("results/07_associated_genes_vulcano_plot.png"), pl2, height = 16, width = 10)
```
